+++ src/main/java/webserver/handler/CreateHandler.java
@@ -1,9 +1,9 @@
 package webserver.handler;
 
 import db.Database;
-import http.HttpRequest;
-import http.HttpResponse;
-import http.HttpStatus;
+import webserver.http.HttpRequest;
+import webserver.http.HttpResponse;
+import webserver.http.HttpStatus;
 import java.util.Map;
 import model.User;
 import org.slf4j.Logger;
@@ -14,8 +14,8 @@ public class CreateHandler implements RequestHandler {
 
     @Override
     public void handle(HttpRequest httpRequest, HttpResponse httpResponse) {
-        Map<String, String> paramMap = httpRequest.getParamMap();
-        User user = User.makeUser(paramMap);
+        Map<String, String> queryMap = httpRequest.getQueryMap();
+        User user = User.makeUser(queryMap);
         Database.addUser(user); // db에 유저 생성
 
         logger.debug("[회원가입] 새로운 유저가 생성되었습니다. {}", user);
