+++ src/main/java/webserver/handler/RequestHandler.java
@@ -1,7 +1,7 @@
 package webserver.handler;
 
-import http.HttpRequest;
-import http.HttpResponse;
+import webserver.http.HttpRequest;
+import webserver.http.HttpResponse;
 
 public interface RequestHandler {
     void handle(HttpRequest httpRequest, HttpResponse httpResponse);
