+++ src/main/java/webserver/http/HttpResponse.java
@@ -1,9 +1,8 @@
-package http;
+package webserver.http;
 
 import java.io.DataOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
-import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import org.slf4j.Logger;
